% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbp.R
\name{gbp}
\alias{gbp}
\title{Finding the target room size in generalized birthday problem}
\arguments{
\item{k}{Integer. k persons shared birthday.}

\item{p}{Integer. Room size to be tested.}

\item{B_l}{Lower bound of tested room size.}

\item{B_u}{Upper bound of upper room size.}

\item{threshold}{A threshold to keep while loop running for binary search.}

\item{print}{Set TRUE if you want to check the result while simulating. Default is FALSE.}

\item{recor}{TRUE if you want to output the room tested in simulation. Default is FALSE.}
}
\value{
Room size
}
\description{
Finding the target room size in generalized birthday problem
}
\examples{
# Simple example with k=2
system.time({
y <- gbp(k=2, p=0.5)
})
y

# You can also track the simulation process
y <- gbp(k=3, p=0.5, print=TRUE)

# Or output the tested room sizes
y <- gbp(k=3, p=0.5, recor=TRUE)
y
}
