\name{GeneralBirthdayProblem-package}
\alias{GeneralBirthdayProblem-package}
\alias{GeneralBirthdayProblem}
\docType{package}
\title{
 Simulate Generalized Birthday Problem
}
\description{
  This package contains a function that find the ideal room size such that the probability of a birthday shared by at least k persons in the room is p.
}
\details{
  
}
\author{
Chang-Yu Chang

Maintainer: Chang-Yu Chang <chang-yu.chang@yale.edu>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
