% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateC.R
\name{simulateC}
\alias{simulateC}
\title{A function for simulation of generalized birthday problem}
\usage{
simulateC(k = 2, room = 23, replicate = 1e+05)
}
\arguments{
\item{k}{Integer. k persons shared birthday.}

\item{room}{Integer. Room size to be tested.}

\item{repliate}{Integer. Replicate size.}
}
\value{
A floating value indicating the probability of at least k persons sharing birthday in a room
}
\description{
A function for simulation of generalized birthday problem
}
\examples{
system.time({
p <- simulateC(k=2, roomSize=23, replicate=1e7)
})
p


}
